Automatically convert python dict to Dict
When setting a dict using the Dict attribute syntax, automatically convert it into a Dict.

This make code like this one valid:

```python
obj = kadet.BaseObj()

obj.root.labels = {"name": "name", "value": "value"}
obj.root.labels.other = "other" # fails without this change as labels would be a dict and not a Dict
```

This change is conservative and only convert the value if type(obj) == dict.
This is to make sure it does not convert a specialised dict (defaultdict, …) if the user chose to use one.